+++ temba/mailroom/client.py
@@ -92,6 +92,21 @@ class MailroomClient:
 
         return self._request("flow/clone", payload)
 
+    def po_export(self, org_id, flow_ids, language, exclude_arguments=False):
+        payload = {
+            "org_id": org_id,
+            "flow_ids": flow_ids,
+            "language": language,
+            "exclude_arguments": exclude_arguments,
+        }
+
+        return self._request("po/export", payload, returns_json=False)
+
+    def po_import(self, org_id, flow_ids, language, po_data):
+        payload = {"org_id": org_id, "flow_ids": flow_ids, "language": language}
+
+        return self._request("po/import", payload, files={"po": po_data})
+
     def sim_start(self, payload):
         return self._request("sim/start", payload)
 
@@ -113,29 +128,35 @@ class MailroomClient:
 
         return self._request("contact/parse_query", payload)
 
-    def _request(self, endpoint, payload=None, post=True):
+    def _request(self, endpoint, payload=None, files=None, post=True, returns_json=True):
         if logger.isEnabledFor(logging.DEBUG):  # pragma: no cover
             logger.debug("=============== %s request ===============" % endpoint)
             logger.debug(json.dumps(payload, indent=2))
             logger.debug("=============== /%s request ===============" % endpoint)
 
+        if files:
+            kwargs = dict(data=payload, files=files)
+        else:
+            kwargs = dict(json=payload)
+
         req_fn = requests.post if post else requests.get
-        response = req_fn("%s/mr/%s" % (self.base_url, endpoint), json=payload, headers=self.headers)
-        resp_json = response.json()
+        response = req_fn("%s/mr/%s" % (self.base_url, endpoint), headers=self.headers, **kwargs)
+
+        return_val = response.json() if returns_json else response.content
 
         if logger.isEnabledFor(logging.DEBUG):  # pragma: no cover
             logger.debug("=============== %s response ===============" % endpoint)
-            logger.debug(json.dumps(resp_json, indent=2))
+            logger.debug(return_val)
             logger.debug("=============== /%s response ===============" % endpoint)
 
         if response.status_code == 422:
-            raise FlowValidationException(endpoint, payload, resp_json)
+            raise FlowValidationException(endpoint, payload, return_val)
         if 400 <= response.status_code < 500:
-            raise MailroomException(endpoint, payload, resp_json)
+            raise MailroomException(endpoint, payload, return_val)
 
         response.raise_for_status()
 
-        return resp_json
+        return return_val
 
 
 def get_client():