+++ temba/mailroom/tests.py
@@ -91,6 +91,33 @@ class MailroomClientTest(TembaTest):
                     },
                 },
             )
+            
+    def test_po_export(self):
+        with patch("requests.post") as mock_post:
+            mock_post.return_value = MockResponse(200, 'msgid "Red"\nmsgstr "Rojo"\n\n')
+            response = get_client().po_export(self.org.id, [123, 234], "spa")
+
+            self.assertEqual(b'msgid "Red"\nmsgstr "Rojo"\n\n', response)
+
+        mock_post.assert_called_once_with(
+            "http://localhost:8090/mr/po/export",
+            headers={"User-Agent": "Temba"},
+            json={"org_id": self.org.id, "flow_ids": [123, 234], "language": "spa", "exclude_arguments": False},
+        )
+
+    def test_po_import(self):
+        with patch("requests.post") as mock_post:
+            mock_post.return_value = MockResponse(200, '{"flows": []}')
+            response = get_client().po_import(self.org.id, [123, 234], "spa", b'msgid "Red"\nmsgstr "Rojo"\n\n')
+
+            self.assertEqual({"flows": []}, response)
+
+        mock_post.assert_called_once_with(
+            "http://localhost:8090/mr/po/import",
+            headers={"User-Agent": "Temba"},
+            data={"org_id": self.org.id, "flow_ids": [123, 234], "language": "spa"},
+            files={"po": b'msgid "Red"\nmsgstr "Rojo"\n\n'},
+        )
 
     def test_parse_query(self):
         with patch("requests.post") as mock_post: