+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+v5.5.12
+----------
+ * Import and export of PO files
+
+v5.5.10
+----------
+ * Bump up the simulator when popped so it fits on more screens
+ * Editor performance improvements
+
 v5.5.8
 ----------
  * Update help text on contact edit dialog